+++ src/components/studyCreate/StudyCreateSelectBox.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { useDispatch } from 'react-redux';
 import DownArrowIcon from "../../assets/icons/communityWrite/purpleDownArrow.svg?react";
-import { setCategoryList } from '../../feautres/study/studyCreateSlice';
+import { setCategoryList } from '../../features/study/studyCreateSlice';
 
 const StudyCreateSelectBox = () => {
     // Redux 관리
