+++ src/components/studyCreate/StudyData.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import StudyCreateLinkEmbed from './StudyCreateLinkEmbed';
-import { setMaterialList } from '../../feautres/study/studyCreateSlice';
+import { setMaterialList } from '../../features/study/studyCreateSlice';
 import { useDispatch } from 'react-redux';
 
 const StudyData = () => {
