--- src/store/anotherSlice.jsx
@@ -1,13 +1,0 @@
-const initialState = {
-  // 초기 상태 정의
-};
-
-const anotherSlice = (state = initialState, action) => {
-  switch (action.type) {
-    // 액션 타입에 따른 상태 변경 로직
-    default:
-      return state;
-  }
-};
-
-export default anotherSlice; // default export 추가
